HitBTC refactored
